+++ resources/views/admin/proveedor/index.blade.php
@@ -74,9 +74,11 @@
                                                     <a href="{{ url('edit-proveedor/'.$proveedor->id) }}" class="btn btn-sm btn-warning" title="Editar">
                                                         <i class="bi bi-pencil-fill"></i>
                                                     </a>
+                                                    @if(Auth::user()->role_as != 1)
                                                     <button type="button" class="btn btn-sm btn-danger" data-bs-toggle="modal" data-bs-target="#deleteModal-{{ $proveedor->id }}" title="Eliminar">
                                                         <i class="bi bi-trash-fill"></i>
                                                     </button>
+                                                    @endif
                                                 </div>
                                             </td>