+++ resources/views/admin/unidad/index.blade.php
@@ -69,9 +69,11 @@
                                                         <a href="{{ url('edit-unidad/'.$unidad->id) }}" class="btn btn-sm btn-warning" title="Editar">
                                                             <i class="bi bi-pencil-fill"></i>
                                                         </a>
+                                                        @if(Auth::user()->role_as != 1)
                                                         <button type="button" class="btn btn-sm btn-danger" data-bs-toggle="modal" data-bs-target="#deleteModal-{{ $unidad->id }}" title="Eliminar">
                                                             <i class="bi bi-trash-fill"></i>
                                                         </button>
+                                                        @endif
                                                     </div>
                                                 </td>
                                                 <td>